v2.7 Fixed json and .bat files
Added microsoft updates to autounattend file.

Changed shell to windows-shell in provisioner

removed copy file of vm guest tools

The packer file now runs and functions all the way to the end with no errors.

Next step is to clean up unused scripts and create a script to hook the server up to AD and have it update with wsus instead of windows updates.